Move GQL resolvers and services under the same package
